feat: create detailComponent
