+++ src/components/MovieList/MovieDetail/index.tsx
@@ -1,0 +1,26 @@
+import React from 'react'
+
+type MovieDetailProps = {
+    item: any
+}
+
+export default function MovieDetail({ item }: MovieDetailProps) {
+    const movieRank = item.rating >= 9 ? 'movieRankGood' : item.rating >= 7 ? 'movieRankSoso' : 'movieRankBad' 
+    const hotIcon = item.rating >= 9 && '🔥'
+    const genres = item.genres.join(', ')
+    const runtime = item.runtime + "min"|| "상영시간 정보없음"
+    const summary = item.summary || "줄거리 없음"      
+    
+    const detail = (
+        <div>
+            <div>{genres}</div>
+            <div>{runtime}</div>
+            <div><span>{hotIcon}</span> 평점 : <span className={movieRank}>{item.rating}</span> / 10</div>
+            <div>{summary}</div>
+        </div>
+    )
+
+    return (
+        <div>{detail}</div>
+    )
+}
